LayoutTests/fast/forms/validation-message-clone-expected.txt

There are two text fields. The first one should have a validation message bubble, and another should have no validation message.




PASS internals.userAgentPartId(lastInOriginalShadow) == "-webkit-validation-bubble" is true
PASS internals.userAgentPartId(lastInClonedShadow) == "-webkit-validation-bubble" is false
